task_template/app/requirements.txt
fastapi
openai
uvicorn
grpcio
grpcio-tools
protobuf
pydantic
langchain
langchain-openai
itsdangerous